requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0